requirements.txt
Flask==2.3.3
gunicorn==20.1.0
python-dotenv==1.0.0
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.0
requests==2.31.0
flask_cors